Moved User.js from helper to lib folder
