+++ index.js
@@ -4,7 +4,7 @@ const AbiBinProvider = require('./lib/AbiBinProvider');
 const Contracts = require('./lib/Contracts');
 const TokenRulesSetup = require('./lib/setup/TokenRules');
 const User = require('./lib/setup/User');
-const UserHelper = require('./lib/helper/User');
+const UserHelper = require('./lib/User');
 const TokenRulesHelper = require('./lib/helper/TokenRules');
 const TokenHolderHelper = require('./lib/helper/TokenHolder');
 const Rules = require('./lib/setup/Rules');
