+++ lib/User.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const AbiBinProvider = require('../AbiBinProvider');
+const AbiBinProvider = require('./AbiBinProvider');
 
 const UserWalletFactoryContractName = 'UserWalletFactory';
 
@@ -8,7 +8,7 @@ const ProxyFactoryContractName = 'ProxyFactory';
 
 const THMasterCopyContractName = 'TokenHolder';
 
-const Utils = require('../../utils/Utils');
+const Utils = require('../utils/Utils');
 
 /**
  * Class is used to create a wallet of an user and configure it.
